Made a number of improvements to globalnav:
When new notifications, make it look like others even though we don't have the number of new items.
For number of messages, we do have that number so just show it and display like others.
For the first four items, always display as fully opaque and visible so they're easy to see