A working Electron desktop app! WIth notifications!
Getting this to build right in the mono repo involved a lot of wackiness. This is because i didnt want to duplicate the notification title, body and url code and moved into the presenters package under NotificationPresenter. And also for this i moved the web navigation utils into the navigation package. But electron cant pull in shared packages from the monorepo when building so they have to be published using yalc and added to the desktop repo this way. Annoying.

But lots of stuff got moved into shared packages across the codebase!