Many improvements to chat room
Use newPostCount to determine how many past and future posts we actually need to load - improving performance.

button to jump to latest actually does so when there are lots of new posts

when loading chat room the new posts now start at the top of the screen instead of below the bottom of the screen.

Load less posts at a time as you scroll so that should be faster

Loading spinner in the middle of the chat room.